+++ src/main/java/jared/common/DukeException.java
@@ -1,3 +1,5 @@
+package jared.common;
+
 public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
