+++ src/main/java/jared/task/Deadline.java
@@ -1,17 +1,19 @@
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.time.LocalTime;
+package jared.task;
 
-public class Event extends Task {
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
+
+public class Deadline extends Task {
     protected LocalDate date;
     protected LocalTime time;
 
-    public Event(String description, LocalDate date) {
+    public Deadline(String description, LocalDate date) {
         super(description);
         this.date = date;
     }
 
-    public Event(String description, LocalDate date, LocalTime time) {
+    public Deadline(String description, LocalDate date, LocalTime time) {
         super(description);
         this.date = date;
         this.time = time;
@@ -19,7 +21,7 @@ public class Event extends Task {
 
     @Override
     public String toString() {
-        String res = "[E]" + super.toString() + " (at: "
+        String res = "[D]" + super.toString() + " (by: "
                 + date.format(DateTimeFormatter.ofPattern("d MMM yyyy"));
         if (time != null) {
             res += " " + time.format(DateTimeFormatter.ofPattern("H:mm"));
@@ -29,7 +31,7 @@ public class Event extends Task {
 
     @Override
     public String saveFormat() {
-        String res = "E _ " + super.saveFormat() + " _ " + this.date;
+        String res = "D _ " + super.saveFormat() + " _ " + this.date;
         if (time != null) {
             res += " _ " + this.time;
         }
