+++ src/main/java/jared/task/Event.java
@@ -1,17 +1,19 @@
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
+package jared.task;
 
-public class Deadline extends Task {
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.time.LocalTime;
+
+public class Event extends Task {
     protected LocalDate date;
     protected LocalTime time;
 
-    public Deadline(String description, LocalDate date) {
+    public Event(String description, LocalDate date) {
         super(description);
         this.date = date;
     }
 
-    public Deadline(String description, LocalDate date, LocalTime time) {
+    public Event(String description, LocalDate date, LocalTime time) {
         super(description);
         this.date = date;
         this.time = time;
@@ -19,7 +21,7 @@ public class Deadline extends Task {
 
     @Override
     public String toString() {
-        String res = "[D]" + super.toString() + " (by: "
+        String res = "[E]" + super.toString() + " (at: "
                 + date.format(DateTimeFormatter.ofPattern("d MMM yyyy"));
         if (time != null) {
             res += " " + time.format(DateTimeFormatter.ofPattern("H:mm"));
@@ -29,7 +31,7 @@ public class Deadline extends Task {
 
     @Override
     public String saveFormat() {
-        String res = "D _ " + super.saveFormat() + " _ " + this.date;
+        String res = "E _ " + super.saveFormat() + " _ " + this.date;
         if (time != null) {
             res += " _ " + this.time;
         }
