+++ src/main/java/jared/task/Task.java
@@ -1,3 +1,5 @@
+package jared.task;
+
 public class Task {
     protected String description;
     protected boolean isDone;
